+++ meshwell/model.py
@@ -4,6 +4,8 @@ from os import cpu_count
 from typing import Dict, Optional, List, Tuple
 import numpy as np
 
+from pathlib import Path
+
 import gmsh
 from meshwell.validation import (
     validate_dimtags,
@@ -147,13 +149,13 @@ class Model:
     def mesh(
         self,
         entities_list: List,
-        resolutions: Optional[Dict] = None,
+        background_remeshing_file: Optional[Path] = None,
         default_characteristic_length: float = 0.5,
         global_scaling: float = 1.0,
         global_2D_algorithm: int = 6,
         global_3D_algorithm: int = 1,
-        filename: Optional[str] = None,
-        verbosity: Optional[int] = 0,
+        filename: Optional[str | Path] = None,
+        verbosity: Optional[int] = 5,
         progress_bars: bool = True,
         interface_delimiter: str = "___",
         boundary_delimiter: str = "None",
@@ -162,12 +164,13 @@ class Model:
         reinitialize: bool = True,
         periodic_entities: List[Tuple[str, str]] = None,
         fuse_entities_by_name: bool = False,
+        optimization_flags: tuple[tuple[str, int]] | None = None,
     ) -> meshio.Mesh:
         """Creates a GMSH mesh with proper physical tagging from a dict of {labels: list( (GMSH entity dimension, GMSH entity tag) )}.
 
         Args:
             entities_list: list of meshwell entities (GMSH_entity, Prism, or PolySurface)
-            resolutions (Dict): Pairs {"physical name": {"resolution": float, "distance": "float}}
+            background_remeshing (Path): path to a .pos file for background remeshing. If not None, is used instead of entity resolutions.
             default_characteristic_length (float): if resolutions is not specified for this physical, will use this value instead
             global_scaling: factor to scale all mesh coordinates by (e.g. 1E-6 to go from um to m)
             global_2D_algorithm: gmsh surface default meshing algorithm, see https://gmsh.info/doc/texinfo/gmsh.html#Mesh-options
@@ -181,13 +184,21 @@ class Model:
             finalize: if True (default), finalizes the GMSH model after execution
             periodic_entities: enforces mesh periodicity between the physical entities
             fuse_entities_by_name: if True, fuses CAD entities sharing the same physical_name
+            optimization_flags: list of (method, niters) for smoothing. See https://gitlab.onelab.info/gmsh/gmsh/blob/gmsh_4_12_2/api/gmsh.py#L2087
 
         Returns:
-            meshio object with mWesh information
+            meshio object with mesh information
         """
+        # Parse filename
+        filename = str(filename)
 
-        resolutions = resolutions if resolutions else {}
-        gmsh.option.setNumber("General.Terminal", verbosity)  # 1 verbose, 0 otherwise
+        # If background mesh, create separate model
+        if background_remeshing_file:
+            # path = os.path.dirname(os.path.abspath(__file__))
+            gmsh.merge(background_remeshing_file)
+            gmsh.model.add("temp")
+
+        gmsh.option.setNumber("General.Terminal", 10)  # 1 verbose, 0 otherwise
         gmsh.option.setNumber(
             "Mesh.CharacteristicLengthMax", default_characteristic_length
         )
@@ -338,24 +349,30 @@ class Model:
                 self.model.occ.remove(entity.dimtags, recursive=True)
 
         # Perform refinement
-        refinement_field_indices = []
-        refinement_max_index = 0
-        for entity in final_entity_list:
-            (
-                refinement_field_indices,
-                refinement_max_index,
-            ) = entity.add_refinement_fields_to_model(
-                refinement_field_indices,
-                refinement_max_index,
-                default_characteristic_length,
-            )
+        if background_remeshing_file is None:
+            # Use entity information
+            refinement_field_indices = []
+            refinement_max_index = 0
+            for entity in final_entity_list:
+                (
+                    refinement_field_indices,
+                    refinement_max_index,
+                ) = entity.add_refinement_fields_to_model(
+                    refinement_field_indices,
+                    refinement_max_index,
+                    default_characteristic_length,
+                )
 
-        # Use the smallest element size overall
-        self.model.mesh.field.add("Min", refinement_max_index)
-        self.model.mesh.field.setNumbers(
-            refinement_max_index, "FieldsList", refinement_field_indices
-        )
-        self.model.mesh.field.setAsBackgroundMesh(refinement_max_index)
+            # Use the smallest element size overall
+            self.model.mesh.field.add("Min", refinement_max_index)
+            self.model.mesh.field.setNumbers(
+                refinement_max_index, "FieldsList", refinement_field_indices
+            )
+            self.model.mesh.field.setAsBackgroundMesh(refinement_max_index)
+        else:
+            bg_field = self.model.mesh.field.add("PostView")
+            self.model.mesh.field.setNumber(bg_field, "ViewIndex", 0)
+            gmsh.model.mesh.field.setAsBackgroundMesh(bg_field)
 
         # Turn off default meshing options
         self.model.mesh.MeshSizeFromPoints = 0
@@ -371,6 +388,12 @@ class Model:
             if global_3D_algorithm == 1 and verbosity:
                 gmsh.logger.start()
             self.model.mesh.generate(max_dim)
+
+            # Mesh smoothing
+            if optimization_flags is not None:
+                for optimization_flag, niter in optimization_flags:
+                    self.model.mesh.optimize(optimization_flag, niter=niter)
+
             if global_3D_algorithm == 1 and verbosity:
                 for line in gmsh.logger.get():
                     if "Optimizing volume " in str(line):