+++ docs/changelog.md
@@ -1,5 +1,24 @@
 # [Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 1.0.7
+
+- filename can be pathlib.Path [commit a755de1](https://github.com/simbilod/meshwell/commit/a755de191140d267f4012ffd9f7b149762281091)
+
+## 1.0.6
+
+Speed improvements:
+- surface extrusion for buffer-free prisms [PR#59](https://github.com/simbilod/meshwell/pull/59)
+- faster interior definition [PR#58](https://github.com/simbilod/meshwell/pull/58)
+
+## 1.0.5
+
+- smoothing options [PR#51](https://github.com/simbilod/meshwell/pull/51)
+
+## 1.0.4
+
+- background remeshing [PR#50](https://github.com/simbilod/meshwell/pull/50)
+- minor fixes
+
 ## 1.0.3
 
 - step output [PR#43](https://github.com/simbilod/meshwell/pull/43)